feat(models): extract sanitization helpers into dedicated module
- Move URL sanitization, invisible char stripping, and scheme validation
  from models/responses.py to new models/sanitization.py
- Reduce coupling in response models and improve testability
- Update tests to import from new sanitization module
- Maintain full backward compatibility with public API
